Created Login/Register pages.
