+++ frontend/index.html
@@ -7,7 +7,7 @@
     <title>Vite + React</title>
   </head>
   <body>
-    <div id="root"></div>
+    <div id="root" class="p-0 m-0"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
