+++ frontend/pages/Login.jsx
@@ -1,0 +1,167 @@
+import React, { useState } from "react";
+import { Link } from "react-router-dom";
+
+function Login() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [remember, setRemember] = useState(false);
+  const [errors, setErrors] = useState({});
+
+  const validate = () => {
+    const newErrors = {};
+    if (!email) newErrors.email = "Email is required";
+    else if (!/^[^@\s]+@[^@\s]+\.[^@\s]+$/.test(email))
+      newErrors.email = "Invalid email";
+    if (!password) newErrors.password = "Password is required";
+    setErrors(newErrors);
+    return Object.keys(newErrors).length === 0;
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (validate()) {
+      alert("Logged in!");
+    }
+  };
+
+  return (
+    <div className="h-screen w-screen flex items-center justify-center bg-gray-900">
+      <div className="bg-gray-800 text-white shadow-lg rounded-lg px-12 py-10 w-full max-w-2xl m-4">
+        <div className="text-center mb-4">
+          <a
+            href="/login"
+            className="d-flex justify-content-center align-items-center gap-2 mb-2 text-decoration-none"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="60"
+              height="60"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              strokeWidth="2"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              className="text-primary"
+            >
+              <path d="m14.5 12.5-8 8a2.119 2.119 0 1 1-3-3l8-8"></path>
+              <path d="m16 16 6-6"></path>
+              <path d="m8 8 6-6"></path>
+              <path d="m9 7 8 8"></path>
+              <path d="m21 11-8-8"></path>
+            </svg>
+            <span className="fs-2 fw-bold text-white">JustBet</span>
+          </a>
+
+          <h2 className="text-2xl font-semibold">Welcome Back</h2>
+          <p className="text-gray-400">Sign in to your account</p>
+        </div>
+
+        {/* Form */}
+        <form onSubmit={handleSubmit} noValidate>
+          <div className="flex flex-col gap-6">
+            {/* Email Row */}
+            <div>
+              <div className="relative">
+                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                  <svg width="20" height="20" fill="none" viewBox="0 0 24 24">
+                    <path
+                      d="M4 4h16v16H4V4zm0 0l8 8 8-8"
+                      stroke="#3B82F6"
+                      strokeWidth="2"
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                    />
+                  </svg>
+                </div>
+                <input
+                  type="email"
+                  className={`w-full pl-10 pr-4 py-2 rounded-md bg-gray-700 text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 ${
+                    errors.email ? "border border-red-500" : ""
+                  }`}
+                  placeholder="Enter your email"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+              </div>
+              {errors.email && (
+                <p className="text-sm text-red-400 mt-1">{errors.email}</p>
+              )}
+            </div>
+
+            {/* Password Row */}
+            <div>
+              <div className="relative">
+                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                  <svg width="20" height="20" fill="none" viewBox="0 0 24 24">
+                    <path
+                      d="M12 17a2 2 0 100-4 2 2 0 000 4zm6-6V9a6 6 0 10-12 0v2a2 2 0 00-2 2v5a2 2 0 002 2h12a2 2 0 002-2v-5a2 2 0 00-2-2z"
+                      stroke="#3B82F6"
+                      strokeWidth="2"
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                    />
+                  </svg>
+                </div>
+                <input
+                  type="password"
+                  className={`w-full pl-10 pr-4 py-2 rounded-md bg-gray-700 text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500 ${
+                    errors.password ? "border border-red-500" : ""
+                  }`}
+                  placeholder="Enter your password"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                />
+              </div>
+              {errors.password && (
+                <p className="text-sm text-red-400 mt-1">{errors.password}</p>
+              )}
+            </div>
+
+            {/* Remember & Forgot */}
+            <div className="flex items-center justify-between">
+              <label className="inline-flex items-center text-lg">
+                <input
+                  type="checkbox"
+                  checked={remember}
+                  onChange={(e) => setRemember(e.target.checked)}
+                  className="form-checkbox h-4 w-4 text-blue-500 mr-2"
+                />
+                <span className="ml-1 text-lg">Remember me</span>
+              </label>
+              <Link
+                to="/forgot-password"
+                className="text-lg text-blue-400 hover:underline"
+              >
+                Forgot password?
+              </Link>
+            </div>
+
+            {/* Submit Button */}
+            <div className="text-center">
+              <button
+                type="submit"
+                className="w-full py-2 px-4 bg-blue-600 hover:bg-blue-700 rounded-lg font-semibold transition"
+              >
+                Sign In
+              </button>
+            </div>
+          </div>
+        </form>
+
+        {/* Footer */}
+        <div className="text-center mt-6 text-lg">
+          Don't have an account?{" "}
+          <Link
+            to="/register"
+            className="text-blue-400 hover:underline font-medium"
+          >
+            Sign up
+          </Link>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Login;
