+++ frontend/src/App.jsx
@@ -3,9 +3,9 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import Login from './Login';
-import Register from './Register';
-import ForgotPassword from './ForgotPassword';
+import Login from '../pages/Login';
+import Register from '../pages/Register';
+
 
 function App() {
   const [count, setCount] = useState(0)
@@ -15,7 +15,6 @@ function App() {
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/forgot-password" element={<ForgotPassword />} />
       </Routes>
     </BrowserRouter>
   )
