+++ frontend/src/index.css
@@ -1,202 +1,1 @@
-:root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
-}
-
-body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-  background: linear-gradient(270deg, #000e2e, #000d23, #011631);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
-}
-
-/* Login Page Styles */
-.login-card {
-  background: #181A20;
-  border-radius: 18px;
-  padding: 10px;
-  box-shadow: 0 4px 32px 0 #000a  ;
-  padding: 2.5rem 2rem;
-  min-width: 370px;
-  max-width: 90vw;
-  border: none;
-}
-.login-title {
-  font-size: 2.1rem;
-  font-weight: bold;
-  color: #fff;
-  margin-top: 1rem;
-}
-.login-subtitle {
-  color: #b0b3c6;
-  font-size: 1rem;
-  margin-bottom: 1.5rem;
-}
-.login-label {
-  color: #fff;
-  font-weight: 500;
-  margin-bottom: 0.3rem;
-}
-.login-input-group {
-  display: flex;
-  align-items: center;
-  background: #23242b;
-  border-radius: 8px;
-  padding: 0.5rem 0.75rem;
-  margin-bottom: 0.5rem;
-}
-.login-input-group svg {
-  margin-right: 0.5rem;
-  color: #3B82F6;
-}
-.login-input {
-  background: transparent;
-  border: none;
-  outline: none;
-  color: #fff;
-  width: 100%;
-  font-size: 1rem;
-}
-.login-input::placeholder {
-  color: #b0b3c6;
-}
-.login-error {
-  color: #ff4d4f;
-  font-size: 0.95rem;
-  margin-top: -0.3rem;
-  margin-bottom: 0.5rem;
-}
-.login-remember {
-  color: #b0b3c6;
-  font-size: 0.98rem;
-}
-.login-link {
-  color: #3B82F6;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.2s;
-}
-.login-link:hover {
-  color: #2563eb;
-}
-.login-btn {
-  background: #2563eb;
-  color: #fff;
-  font-weight: bold;
-  border: none;
-  border-radius: 8px;
-  width: 100%;
-  padding: 0.7rem 0;
-  font-size: 1.1rem;
-  margin-bottom: 0.5rem;
-  margin-top: 0.5rem;
-  transition: background 0.2s;
-}
-.login-btn:hover {
-  background: #1d4ed8;
-}
-.login-footer {
-  text-align: center;
-  margin-top: 1.2rem;
-  color: #b0b3c6;
-  font-size: 1rem;
-}
-.login-footer .login-link {
-  font-weight: bold;
-}
-
-@media (max-width: 600px) {
-  .login-card {
-    min-width: unset;
-    width: 100vw;
-    max-width: 98vw;
-    padding: 1.2rem 0.5rem;
-    border-radius: 10px;
-  }
-  .login-title {
-    font-size: 1.4rem;
-  }
-  .login-subtitle {
-    font-size: 0.95rem;
-  }
-  .login-footer {
-    font-size: 0.95rem;
-  }
-  .login-input-group {
-    padding: 0.4rem 0.5rem;
-  }
-  .login-btn {
-    font-size: 1rem;
-    padding: 0.6rem 0;
-  }
-}
-
-@media (max-width: 400px) {
-  .login-card {
-    padding: 0.5rem 0.1rem;
-  }
-  .login-title {
-    font-size: 1.1rem;
-  }
-}
+@import "tailwindcss";
